Initial working version of the program.
In order to keep it simple, only a single thread is used for both fetching all
the data (network I/O) and also aggregating the results (CPU/RAM).
If the data size grows considerably, we should introduce asyncio and download
pages in parallel to speed up the process.

Configurable timeouts are used when communicating with the REST API.
Retries and exponential backoff would be a nice improvement to add.

Added some tests for the core components.
Coverage should be improved.